+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 - Add graphql request duration reporting via volley - ds300
 - Fixes `Contact Gallery` button showing up when artwork auction is closed - sepans
+- Updates CommercialInformation component to properly handle live auctions - sweir27
 
 ### 1.17.8